safari-600.7-branch/LayoutTests/fast/dom/Window/invalid-protocol-expected.txt

Test URL protocol setter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS location.protocol = '' threw exception Error: SyntaxError: DOM Exception 12.
PASS location.protocol = ':' threw exception Error: SyntaxError: DOM Exception 12.
PASS location.protocol = 'é' threw exception Error: SyntaxError: DOM Exception 12.
PASS location.protocol = '[' threw exception Error: SyntaxError: DOM Exception 12.
PASS location.protocol = '0' threw exception Error: SyntaxError: DOM Exception 12.
PASS a.protocol is 'http:'
PASS a.href is 'https://www.apple.com/'
PASS a.href is 'http://www.apple.com/'
PASS a.href is 'https://www.apple.com/'
PASS successfullyParsed is true

TEST COMPLETE
